+++ src/Model/User/UserParams.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Fire\Model\User;
+
+use Fire\Foundation\Params;
+
+class UserParams extends Params
+{
+
+}