+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased][unreleased]
 
+## [2.0.3] - 2015-09-09
+### Changed
+- Changed repository `newParams` method to public visibility
+
+### Added
+- Added User params builder
+
 ## [2.0.2] - 2015-09-08
 ### Added
 - Added `limitWords` and `limitChars` helper functions
@@ -14,4 +21,6 @@
 ### Added
 - Initial public release
 
+[2.0.3]: https://github.com/emrl/fire/compare/2.0.2...2.0.3
+[2.0.2]: https://github.com/emrl/fire/compare/2.0.1...2.0.2
 [2.0.1]: https://github.com/emrl/fire/compare/2.0.0...2.0.1